Simplify `material-3-codelab`
