+++ material-3-codelab/step_05/lib/main.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/material.dart';
 
 import 'models/data.dart' as data;
-import 'models/models.dart';
+import 'models/model.dart';
 import 'widgets/email_list_view.dart';
 
 void main() {
