+++ material-3-codelab/step_05/lib/models/model.dart
@@ -2,7 +2,13 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'models.dart';
+class Attachment {
+  const Attachment({
+    required this.url,
+  });
+
+  final String url;
+}
 
 class Email {
   const Email({
@@ -20,4 +26,27 @@ class Email {
   final String content;
   final List<Attachment> attachments;
   final double replies;
+}
+
+class Name {
+  const Name({
+    required this.first,
+    required this.last,
+  });
+
+  final String first;
+  final String last;
+  String get fullName => '$first $last';
+}
+
+class User {
+  const User({
+    required this.name,
+    required this.avatarUrl,
+    required this.lastActive,
+  });
+
+  final Name name;
+  final String avatarUrl;
+  final DateTime lastActive;
 }
