--- material-3-codelab/step_06/lib/models/name.dart
@@ -1,14 +1,0 @@
-// Copyright 2022 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-class Name {
-  const Name({
-    required this.first,
-    required this.last,
-  });
-
-  final String first;
-  final String last;
-  String get fullName => '$first $last';
-}
