+++ material-3-codelab/step_06/lib/widgets/search_bar.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter/material.dart';
 
-import '../models/user.dart';
+import '../models/model.dart';
 
 class SearchBar extends StatelessWidget {
   const SearchBar({
