--- material-3-codelab/step_07/lib/models/models.dart
@@ -1,8 +1,0 @@
-// Copyright 2022 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-export 'attachment.dart';
-export 'email.dart';
-export 'name.dart';
-export 'user.dart';
