--- material-3-codelab/step_09/lib/models/attachment.dart
@@ -1,11 +1,0 @@
-// Copyright 2022 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-class Attachment {
-  const Attachment({
-    required this.url,
-  });
-
-  final String url;
-}
