+++ material-3-codelab/step_09/lib/models/data.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'models.dart';
+import 'model.dart';
 
 final User user_0 = User(
     name: const Name(first: 'Me', last: ''),
