--- material-3-codelab/step_09/lib/models/user.dart
@@ -1,17 +1,0 @@
-// Copyright 2022 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'models.dart';
-
-class User {
-  const User({
-    required this.name,
-    required this.avatarUrl,
-    required this.lastActive,
-  });
-
-  final Name name;
-  final String avatarUrl;
-  final DateTime lastActive;
-}
