+++ material-3-codelab/step_09/lib/widgets/email_list_view.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/material.dart';
 
 import '../models/data.dart' as data;
-import '../models/user.dart';
+import '../models/model.dart';
 import 'email_widget.dart';
 import 'search_bar.dart';
