+++ material-3-codelab/step_09/lib/widgets/email_widget.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import '../models/models.dart';
+import '../models/model.dart';
 import 'star_button.dart';
 
 enum EmailType {
